Introduce Parameter object
In this commit, we implement a parameter object based on a singleton design pattern.
This object is in charge of reading from the file and storing the user input
parameters within its state. Due to the singleton design pattern only one
instance of this class exist which can be shared with other classes for their
respective accesses.

One of the difficulties would be figuring out how to unit test this class and
all other classes using this class since mocking a static membered class is difficult.