Arch for IO operations (write only)
In this commit, we put in place the necessary architecture for
an IO manager that will read and write files in the application.
It is based on factory method and an abstract factory class.
Currently only vtk structured grid writing is supported.
This version does not compile due to an error.